add docker memory limit and revervation
Memory Limit: A strict upper limit to the amount of memory made available to a container.
Memory Reservation: This should be set as the bare minimum amount of memory that an application needs to run properly. So it doesn’t crash or misbehave when the system is trying to reclaim some of the memory.